requirements.txt
fastapi
uvicorn
streamlit
PyPDF2
faiss-cpu
sentence-transformers
